chore(oxc-minify) use `minifySync` instead of async (#839)
* feat: use `minifySync` instead of async for `oxc-minify`

Using the sync version because the async version spawns a real thread,
which is slow for single threaded js benchmarks.

The purpose of this PR is to observe performance difference between the
async and sync version, feel free to reject if it does not meet the
requirements of this project.

* wip

* wip

* chore: updated benchmarks

---------

Co-authored-by: Boshen <boshenc@gmail.com>
Co-authored-by: GitHub <github-actions@github.com>